--- lib/args/border.dart
@@ -1,16 +1,0 @@
-import 'package:flutter/widgets.dart' show Border;
-
-import 'package:pdf/widgets.dart' as pw show Border;
-
-import 'border_side.dart';
-
-
-extension BorderConverter on Border {
-  pw.Border toPdfBorder() => pw.Border(
-    top: top.toPdfBorderSide(),
-    right: right.toPdfBorderSide(),
-    bottom: bottom.toPdfBorderSide(),
-    left: left.toPdfBorderSide(),
-  );
-}
-
