+++ lib/widgets/button.dart
@@ -1,12 +1,13 @@
-import 'package:flutter/material.dart' show ButtonStyleButton, EdgeInsets;
+import 'package:flutter/material.dart' show ButtonStyleButton, EdgeInsets,  BeveledRectangleBorder, BorderRadius, CircleBorder, ContinuousRectangleBorder, OutlinedBorder, RoundedRectangleBorder, StadiumBorder, debugPrint;
 
 import 'package:pdf/pdf.dart' show PdfColor, PdfColors;
-import 'package:pdf/widgets.dart' as pw show FlatButton, Widget, BoxDecoration, Border, BoxShape, EdgeInsets;
+import 'package:pdf/widgets.dart' as pw show FlatButton, Widget, BoxDecoration, Border, BoxShape, EdgeInsets, BorderRadius;
 
 import '/args/border_side.dart';
-import '/args/outlined_border.dart';
 import '/args/color.dart';
 import '/args/edge_insets.dart';
+import '/args/border_radius.dart';
+
 
 
 extension ButtonConverter on ButtonStyleButton {
@@ -28,4 +29,47 @@ extension ButtonConverter on ButtonStyleButton {
         ?? const pw.EdgeInsets.all(4.0),
     child: child,
   );
+}
+
+extension OutlinedBorderConverter on OutlinedBorder {
+  pw.BoxShape toPdfBoxShape() {
+    switch (runtimeType) {
+      case CircleBorder:
+        return pw.BoxShape.circle;
+      case StadiumBorder:
+        return pw.BoxShape.circle;
+      case RoundedRectangleBorder:
+        return pw.BoxShape.rectangle;
+      case BeveledRectangleBorder:
+        return pw.BoxShape.rectangle;
+      case ContinuousRectangleBorder:
+        return pw.BoxShape.rectangle;
+      // unsupported by pdf package:
+      // - StadiumBorder
+      // - RoundedRectangleBorder
+      // - BeveledRectangleBorder
+      // - ContinuousRectangleBorder
+      default:
+        debugPrint('Unsupported OutlinedBorder: $this; defaulting to rectangle border shape');
+        return pw.BoxShape.rectangle;
+    }
+  }
+
+  pw.BorderRadius toPdfBorderRadius() {
+    switch (runtimeType) {
+      case CircleBorder:
+        return pw.BorderRadius.circular(1000);
+      case StadiumBorder:
+        return pw.BorderRadius.circular(10);
+      case RoundedRectangleBorder:
+        return ((this as RoundedRectangleBorder).borderRadius as BorderRadius).toPdfBorderRadius();
+      case BeveledRectangleBorder:
+        return ((this as BeveledRectangleBorder).borderRadius as BorderRadius).toPdfBorderRadius();
+      case ContinuousRectangleBorder:
+        return ((this as ContinuousRectangleBorder).borderRadius as BorderRadius).toPdfBorderRadius();
+      default:
+        debugPrint('Unsupported OutlinedBorder: $this; defaulting to circular border radius');
+        return pw.BorderRadius.circular(1000);
+    }
+  }
 }
