+++ lib/widgets/stack.dart
@@ -1,9 +1,8 @@
-import 'package:flutter/widgets.dart' show Alignment, AlignmentDirectional, Stack;
+import 'package:flutter/widgets.dart' show Alignment, AlignmentDirectional, Stack, StackFit;
 
-import 'package:pdf/widgets.dart' as pw show Stack, Widget;
+import 'package:pdf/widgets.dart' as pw show Stack, Widget, StackFit;
 
 import '/args/alignment.dart';
-import '/args/stack_fit.dart';
 
 
 extension StackConverter on Stack {
@@ -14,4 +13,17 @@ extension StackConverter on Stack {
     fit: fit.toPdfStackFit(),
     children: children,
   );
+}
+
+extension StackFitConverter on StackFit {
+  pw.StackFit toPdfStackFit() {
+    switch (this) {
+      case StackFit.loose:
+        return pw.StackFit.loose;
+      case StackFit.expand:
+        return pw.StackFit.expand;
+      case StackFit.passthrough:
+        return pw.StackFit.passthrough;
+    }
+  }
 }
