+++ lib/widgets/text.dart
@@ -1,11 +1,10 @@
-import 'package:flutter/widgets.dart' show Text;
+import 'package:flutter/widgets.dart' show Text, TextOverflow;
 
-import 'package:pdf/widgets.dart' as pw show Text;
+import 'package:pdf/widgets.dart' as pw show Text, TextOverflow;
 
 import '/args/text_style.dart';
 import '/args/text_align.dart';
 import '/args/text_direction.dart';
-import '/args/text_overflow.dart';
 
 
 extension TextConverter on Text {
@@ -18,4 +17,20 @@ extension TextConverter on Text {
     textScaleFactor: textScaleFactor ?? 1.0,
     style: style?.toPdfTextStyle(),
   );
+}
+
+extension TextOverlfowConverter on TextOverflow {
+  pw.TextOverflow toPdfTextOverflow() {
+    switch (this) {
+      case TextOverflow.clip:
+        return pw.TextOverflow.clip;
+      case TextOverflow.visible:
+        return pw.TextOverflow.visible;
+      // not supported in pdf package:
+      // - TextOverflow.ellipsis
+      // - TextOverflow.fade
+      default:
+        return pw.TextOverflow.span;
+    }
+  }
 }
